+++ CHANGELOG.md
@@ -4,6 +4,7 @@ Changelog
 *Also see [Tools and installer changelog](https://github.com/clojure/brew-install/blob/1.10.3/CHANGELOG.md)*
 
 * next
+  * TDEPS-219 Fixed `-X:deps find-versions` doesn't find Maven versions that start with a letter
   * Fix bad invocation of deps/prep-libs! in clojure.tools.cli.api/prep
 * 0.14.1222 on Aug 5, 2022
   * TDEPS-228 Add support for auto inferred Sourcehut git urls